requirements.txt
fastapi>=0.103.0
pydantic>=2.0
gradio  # latest
requests==2.31.0
#pydantic==1.10.12 older edition
python-dotenv==1.0.0
# aiohttp==3.8.4 required earlier with streamlit perhaps?
pytest==7.4.0
gunicorn==20.1.0



# GCP Integration - Cloud Logging
google-cloud-logging==3.5.0

# Vertex AI Integration (Phase 1 - Real Gemini Models)
google-cloud-aiplatform>=1.35.0
vertexai>=0.30.0
